+++ lib/src/clients/clients.dart
@@ -1,5 +1,6 @@
 export 'converters/converter.dart';
 export 'interceptors/interceptors.dart';
+export 'handle.dart';
 export 'request.dart';
 export 'rest.dart';
 export 'wrapper.dart';