+++ lib/src/clients/converters/converter.dart
@@ -1,55 +1,33 @@
 import 'dart:async';
 
 import 'package:http/http.dart';
-import 'package:meta/meta.dart';
 
 import '../wrapper.dart';
 import 'request.dart';
 import 'response.dart';
 
+export 'exception.dart';
 export 'request.dart';
 export 'response.dart';
 
 /// {@category Clients}
-class ConverterClient extends WrapperClient {
+class ConverterClient extends WrapperClient
+    with RequestConverterMixin, ResponseConverterMixin {
+  @override
   final Iterable<RequestConverterCallback>? requestConverters;
+  @override
   final Iterable<ResponseConverterCallback>? responseConverters;
 
   ConverterClient(
-    Client client, {
+    super.client, {
     this.requestConverters,
     this.responseConverters,
-  }) : super(client);
-
-  @protected
-  FutureOr<BaseRequest> onConvertRequest(BaseRequest request) async {
-    final converters = requestConverters;
-    if (converters == null || converters.isEmpty) return request;
-    BaseRequest modifiedRequest = request;
-
-    for (final converter in converters) {
-      modifiedRequest = await converter(modifiedRequest);
-    }
-
-    return modifiedRequest;
-  }
-
-  @protected
-  FutureOr<StreamedResponse> onConvertResponse(StreamedResponse request) async {
-    final converters = responseConverters;
-    if (converters == null || converters.isEmpty) return request;
-    StreamedResponse modifiedResponse = request;
-
-    for (final converter in converters) {
-      modifiedResponse = await converter(modifiedResponse);
-    }
-
-    return modifiedResponse;
-  }
+  });
 
   @override
   Future<StreamedResponse> send(BaseRequest request) async {
-    final response = await inner.send(await onConvertRequest(request));
+    final modifiedRequest = await onConvertRequest(request);
+    final response = await inner.send(modifiedRequest);
     final modifiedResponse = await onConvertResponse(response);
     return modifiedResponse;
   }