+++ lib/src/clients/converters/request.dart
@@ -4,34 +4,63 @@ import 'package:http/http.dart';
 import 'package:meta/meta.dart';
 
 import '../wrapper.dart';
+import 'exception.dart';
 
 typedef RequestConverterCallback = FutureOr<BaseRequest> Function(
   BaseRequest,
 );
 
-/// {@category Clients}
-class RequestConverterClient extends WrapperClient {
-  final Iterable<RequestConverterCallback> converters;
+class RequestConverterException extends ConverterException {
+  const RequestConverterException(
+    super.message, {
+    super.uri,
+    super.innerException,
+    super.innerStackTrace,
+  });
+}
 
-  RequestConverterClient(
-    Client client,
-    this.converters,
-  ) : super(client);
+mixin RequestConverterMixin {
+  @protected
+  Iterable<RequestConverterCallback>? get requestConverters;
 
   @protected
   FutureOr<BaseRequest> onConvertRequest(BaseRequest request) async {
-    if (converters.isEmpty) return request;
+    final requestConverters = this.requestConverters;
+    if (requestConverters == null || requestConverters.isEmpty) return request;
     BaseRequest modifiedRequest = request;
 
-    for (final converter in converters) {
-      modifiedRequest = await converter(modifiedRequest);
+    for (final converter in requestConverters) {
+      try {
+        modifiedRequest = await converter(modifiedRequest);
+      } catch (e, s) {
+        throw RequestConverterException(
+          'Request Converter failed due to an error',
+          uri: request.url,
+          innerException: e,
+          innerStackTrace: s,
+        );
+      }
     }
 
     return modifiedRequest;
   }
+}
+
+/// {@category Clients}
+class RequestConverterClient extends WrapperClient with RequestConverterMixin {
+  RequestConverterClient(
+    super.client,
+    this.converters,
+  );
+
+  final Iterable<RequestConverterCallback> converters;
+
+  @override
+  Iterable<RequestConverterCallback> get requestConverters => converters;
 
   @override
   Future<StreamedResponse> send(BaseRequest request) async {
-    return inner.send(await onConvertRequest(request));
+    final modifiedRequest = await onConvertRequest(request);
+    return inner.send(modifiedRequest);
   }
 }