+++ test/client/client_conformance_test.dart
@@ -12,6 +12,9 @@ void main() {
   group('RequestClient', () {
     testAll(() => RequestClient(http.Client()));
   });
+  group('HandleClient', () {
+    testAll(() => HandleClient(http.Client()));
+  });
   group('InterceptorClient', () {
     testAll(
       () => InterceptorClient(